PERF/COMPAT: define platform int to np.intp
AFAIK this only affects 64 bit python on Windows.    `numpy` wants an
`np.intp` (i8 on Windows) as a indexer for `take`, but pandas defines
a "platform int" as a `np.int_` (i4 on Windows).  This hits
performance twice, because we often start with i8, cast to i4, then
numpy will cast back to i8 in its `take`.

Author: Chris <cbartak@gmail.com>

Closes #13972 from chris-b1/platform-int and squashes the following commits:
Closes #3033

322b11a [Chris] lint fixup
fc80938 [Chris] adjust for 32bit
84f38b2 [Chris] adjust test for platform independence
3ced5d5 [Chris] PERF/COMPAT: define platform int to np.intp

+++ pandas/hashtable.pyx
@@ -64,10 +64,10 @@ cdef class Factorizer:
         mask = (labels == na_sentinel)
         # sort on
         if sort:
-            if labels.dtype != np.int_:
-                labels = labels.astype(np.int_)
+            if labels.dtype != np.intp:
+                labels = labels.astype(np.intp)
             sorter = self.uniques.to_array().argsort()
-            reverse_indexer = np.empty(len(sorter), dtype=np.int_)
+            reverse_indexer = np.empty(len(sorter), dtype=np.intp)
             reverse_indexer.put(sorter, np.arange(len(sorter)))
             labels = reverse_indexer.take(labels, mode='clip')
             labels[mask] = na_sentinel
@@ -100,11 +100,11 @@ cdef class Int64Factorizer:
 
         # sort on
         if sort:
-            if labels.dtype != np.int_:
-                labels = labels.astype(np.int_)
+            if labels.dtype != np.intp:
+                labels = labels.astype(np.intp)
 
             sorter = self.uniques.to_array().argsort()
-            reverse_indexer = np.empty(len(sorter), dtype=np.int_)
+            reverse_indexer = np.empty(len(sorter), dtype=np.intp)
             reverse_indexer.put(sorter, np.arange(len(sorter)))
 
             labels = reverse_indexer.take(labels)
